+++ src/main/java/prismatic/shards/stellarity/datagen/book/EndonomiconBookProvider.java
@@ -2,6 +2,7 @@ package prismatic.shards.stellarity.datagen.book;
 
 import com.klikli_dev.modonomicon.api.datagen.ModonomiconLanguageProvider;
 import com.klikli_dev.modonomicon.api.datagen.SingleBookSubProvider;
+import com.klikli_dev.modonomicon.api.datagen.book.BookModel;
 import prismatic.shards.stellarity.Stellarity;
 import prismatic.shards.stellarity.datagen.book.category.ArmorsCategory;
 
@@ -13,6 +14,13 @@ public class EndonomiconBookProvider extends SingleBookSubProvider {
 	@Override
 	protected void registerDefaultMacros() {
 
+	}
+
+	@Override
+	protected BookModel additionalSetup(BookModel book) {
+		super.additionalSetup(book);
+		book.withGenerateBookItem(false);
+		return book;
 	}
 
 	@Override