+++ src/main/java/prismatic/shards/stellarity/datagen/dynamic/PlacedFeatureProvider.java
@@ -6,21 +6,20 @@ import net.minecraft.core.HolderGetter;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstrapContext;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.levelgen.Heightmap;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
-import net.minecraft.world.level.levelgen.placement.*;
+import net.minecraft.world.level.levelgen.placement.PlacedFeature;
 import prismatic.shards.stellarity.Stellarity;
 import prismatic.shards.stellarity.key.StellarityConfiguredFeatures;
 
 import java.util.List;
-import java.util.function.Function;
 
-import static net.minecraft.world.level.block.Blocks.*;
+import static net.minecraft.world.level.block.Blocks.AIR;
 import static prismatic.shards.stellarity.key.StellarityPlacedFeatures.*;
-import static prismatic.shards.stellarity.util.ValueUtil.*;
+import static prismatic.shards.stellarity.util.ValueUtil.num;
+import static prismatic.shards.stellarity.util.ValueUtil.weighted;
 import static prismatic.shards.stellarity.util.WorldgenUtil.*;